# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Valet Parking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts for an event
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Open a new shift (admin or owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftBody"
      responses:
        "201":
          description: Created shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{shiftId}:
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update shift (close, modify times) - admin or owner
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShiftBody"
      responses:
        "200":
          description: Updated shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List active valet tickets for an event/shift
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
        - name: shiftId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, in_transit, delivered, relocated]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValetTicket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new valet ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValetTicket"
        "409":
          description: Duplicate valet number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tickets/history:
    get:
      operationId: listTicketHistory
      tags: [tickets]
      summary: Full ticket history for an event (owner only)
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValetTicket"
  /tickets/{ticketId}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID (also used for search by valet number)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValetTicket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket status (search, in_transit, deliver, relocate)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValetTicket"
  /tickets/search/{valetNumber}:
    get:
      operationId: searchTicketByValetNumber
      tags: [tickets]
      summary: Search active ticket by valet number
      parameters:
        - name: valetNumber
          in: path
          required: true
          schema:
            type: string
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValetTicket"
        "404":
          description: Not found
  /parking-locations:
    get:
      operationId: listParkingLocations
      tags: [parking]
      summary: List all parking locations with availability
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Parking locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingLocation"
    post:
      operationId: createParkingLocation
      tags: [parking]
      summary: Create parking location (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParkingLocationBody"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingLocation"
  /parking-locations/{locationId}:
    patch:
      operationId: updateParkingLocation
      tags: [parking]
      summary: Update parking location (owner only)
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParkingLocationBody"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingLocation"
    delete:
      operationId: deleteParkingLocation
      tags: [parking]
      summary: Delete parking location (owner only)
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        label:
          type: string
        openedBy:
          type: string
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        isOpen:
          type: boolean
      required: [id, eventId, label, openedBy, openedAt, isOpen]
    CreateShiftBody:
      type: object
      properties:
        eventId:
          type: integer
        label:
          type: string
        openedBy:
          type: string
      required: [eventId, label, openedBy]
    UpdateShiftBody:
      type: object
      properties:
        label:
          type: string
        closedAt:
          type: string
          format: date-time
        isOpen:
          type: boolean
    ValetTicket:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        shiftId:
          type: integer
        valetNumber:
          type: string
        driverName:
          type: string
        keyLocation:
          type: string
          enum: [drawer, board, not_found, with_owner]
        parkingLocationId:
          type: integer
          nullable: true
        parkingLocationName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, in_transit, delivered, relocated]
        vehicleDamages:
          type: array
          items:
            type: string
            enum: [front_door, rear_bumper, front_bumper, rear_light, front_light, hood]
        notes:
          type: string
          nullable: true
        relocatedToLocationId:
          type: integer
          nullable: true
        relocatedToLocationName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
      required: [id, eventId, shiftId, valetNumber, driverName, keyLocation, status, vehicleDamages, createdAt]
    CreateTicketBody:
      type: object
      properties:
        eventId:
          type: integer
        shiftId:
          type: integer
        valetNumber:
          type: string
        driverName:
          type: string
        keyLocation:
          type: string
          enum: [drawer, board, not_found, with_owner]
        parkingLocationId:
          type: integer
        vehicleDamages:
          type: array
          items:
            type: string
            enum: [front_door, rear_bumper, front_bumper, rear_light, front_light, hood]
        notes:
          type: string
      required: [eventId, shiftId, valetNumber, driverName, keyLocation, parkingLocationId]
    UpdateTicketBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, in_transit, delivered, relocated]
        keyLocation:
          type: string
          enum: [drawer, board, not_found, with_owner]
        parkingLocationId:
          type: integer
        relocatedToLocationId:
          type: integer
        vehicleDamages:
          type: array
          items:
            type: string
            enum: [front_door, rear_bumper, front_bumper, rear_light, front_light, hood]
        notes:
          type: string
    ParkingLocation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        currentCount:
          type: integer
        colorHex:
          type: string
        isActive:
          type: boolean
      required: [id, name, capacity, currentCount, colorHex, isActive]
    CreateParkingLocationBody:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        colorHex:
          type: string
      required: [name, capacity, colorHex]
    UpdateParkingLocationBody:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        colorHex:
          type: string
        isActive:
          type: boolean
...